zamp_public_workflow_sdk/temporal/codec/storage_client.py
import asyncio
from abc import ABC, abstractmethod

from google.cloud import storage


class StorageClient(ABC):
    @abstractmethod
    async def upload_file(self, object_name: str, data: bytes | str) -> None:
        pass

    @abstractmethod
    async def get_file(self, object_name: str) -> bytes:
        pass


class GCSClient(StorageClient):
    def __init__(self, project_id: str, bucket_name: str):
        self.client = storage.Client(project=project_id)
        self.bucket = self.client.bucket(bucket_name)

    async def upload_file(self, object_name: str, data: bytes | str) -> None:
        await asyncio.to_thread(self.bucket.blob(object_name).upload_from_string, data)

    async def get_file(self, object_name: str) -> bytes:
        return await asyncio.to_thread(self.bucket.blob(object_name).download_as_bytes)
